requirements.txt
Django==4.2.16
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow==10.3.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
bcrypt==5.0.0
